create branches for the platform-specific packages
Keeping each package on its own branch, rather than using directories
on a single branch, makes it possible to download the package for one
platform without downloading all of them.

Adding the branches as submodules effectively recreates the old setup,
but with the added benefits of branches.

Each platform-specific branch has been created at the `clean-start` tag.

The planned platforms are:

  - x86_64-linux
  - x86_64-macosx
  - aarch64-macosx
  - i386-win32
  - x86_64-win32

The uniform naming convention is based on `normalize-platform` from:

https://github.com/mflatt/raco-cross/blob/dbb94f2/private/cross/platform.rkt#L61-L85