wip: avoid having to build everything on Darwin
Add a package for Darwin systems to include all Apple hosts and
only Apple hosts. Building for Apple hosts requires a Darwin build
environment, but hopefully this is a step toward not having to build
everything else there. In particular, the Darwin build environment
does not seem to have binary substitutes available for many cross
toolchains targeting non-Apple hosts, so a great deal of time was
consumed building mingw and the like.